Add a command-line option to not play back the song
Not too happy about having to hardcode the size of the sample buffer like this,
but well, that's what the audio backends do as well. Also, it's rather
pointless to support variable sampling rates. This either requires each
hardware to be designed for a variable sample rate, which we don't know, or
manual resampling, which should totally not be our problem.