Refactor format-specific Corlett library loading into a separate function
Unfortunately, PSF is the odd one out here in that it requires the main file
to be patched above the first library, but *not* above the remaining ones.
Not that it would ruin the callback system I'm planning here, it just makes it
a bit awkward for PSF in particular.