PSF: Enforce the intended initialization order entirely within psf_lib()
Of course, this requires all intermediate files and corlett_t structures to not
be freed immediately after the call to psf_lib(), but we can totally define the
callback system to work like that. In the meantime though, I'll just be piggy
and don't free anything. psf_start() already is confusing enough as it is.