mididump: Write the analyzed beat distance
This just sets the MIDI time division to the beat distance in samples and the
MIDI tempo to the same value converted to beats per minute. Without doing any
time value scaling, the 16-bit size of the MIDI time division field effectively
limits us to a minimum tempo of 40.37 BPM, which corresponds to 32,767 samples.