PSF2: Make DMA7 reads and writes actually work
Wow, that code needs some serious compression-oriented cleanup… or I better
just outright replace it with https://github.com/kode54/Highly_Experimental.
But that's beyond the scope of this bugfix. Anyway, the Final Fantasy X rip by
nenolod now plays correctly after the regression in the previous commit, and
thus, issue #1 is fixed. Thanks to kode54 for the tip.