Debug: On Windows, render the debug window in a separate thread
This at least keeps the music playing while the debug window is being moved or
resized. Sadly, we can't get the window to update during that time as well.
This would require extensive rewrites of GLFW, or a completely new Win32
platform layer written from scratch, which really isn't worth it at this point.