Linux: Fall back on not playing anything if /dev/dsp is not present
Also, as pointed out in #3, macOS doesn't have `sys/soundcard.h`, so
it's about time I did something about this. Unfortunately, it's up to
m1sdr_TimeCheck() to actually request a certain number of samples to be
rendered, which complicates the implementation of a "null audio device"
way more than it should. So, a NOPLAY #define is indeed the best option
here, since that can also remove the -p/--noplay option, analogous to
defining NOGUI.